Refactor CandidateList and ScanningCard to integrate GroupsModal and update modal handling; remove deprecated annotations modal logic.
