+++ src/scanning/scanningSession/components/ScanningCard/ScanningCard.jsx
@@ -3,27 +3,20 @@ import { THUMBNAIL_TYPES } from "../../../../sources/sources.lib.js";
 import { Thumbnail } from "../../../../sources/components/Thumbnail/Thumbnail.jsx";
 import { PinnedAnnotations } from "../../../../sources/components/PinnedAnnotations/PinnedAnnotations.jsx";
 import { PhotometryChart } from "../../../../sources/components/PhotometryChart/PhotometryChart.jsx";
-import { memo, useState } from "react";
+import { memo, useRef } from "react";
 import { ScanningCardSkeleton } from "./ScanningCardSkeleton.jsx";
-import {
-  IonButton, IonButtons,
-  IonChip,
-  IonContent,
-  IonHeader,
-  IonModal,
-  IonTitle,
-  IonToolbar
-} from "@ionic/react";
+import { IonChip } from "@ionic/react";
 import { SourceInfo } from "../../../../sources/components/SourceInfo/SourceInfo.jsx";
 import { FollowupRequests } from "../../../../sources/components/FollowupRequests/FollowupRequests.jsx";
 import { Comments } from "../../../../sources/components/Comments/Comments.jsx";
 import { SpectraList } from "../../../../sources/components/Spectra/SpectraList.jsx";
+import { GroupsModal } from "../../../../sources/components/GroupsModal/GroupsModal.jsx";
+import { AnnotationsViewerModal } from "../../../../sources/components/PinnedAnnotations/AnnotationsViewerModal.jsx";
 
 /**
  * Scanning card component
  * @param {Object} props
  * @param {import("../../../scanning.lib.js").Candidate} props.candidate
- * @param {React.MutableRefObject<any>} props.modal
  * @param {number} props.currentIndex
  * @param {number} props.nbCandidates
  * @param {boolean} props.isInView
@@ -32,13 +25,16 @@ import { SpectraList } from "../../../../sources/components/Spectra/SpectraList.
  */
 const ScanningCardBase = ({
   candidate,
-  modal,
   currentIndex,
   nbCandidates,
   isInView,
   pinnedAnnotations,
 }) => {
-  const [showGroupsSaveTo, setShowGroupsSaveTo] = useState(false);
+  /** @type {React.MutableRefObject<any>} */
+  const groupsModal = useRef(null);
+  /** @type {React.MutableRefObject<any>} */
+  const annotationsModal = useRef(null);
+
   return (
     <div className="scanning-card-container">
       <div
@@ -51,7 +47,7 @@ const ScanningCardBase = ({
             className="is-saved"
             color={candidate.is_source ? "primary" : "secondary"}
             onClick={() => {
-              if (candidate.is_source) setShowGroupsSaveTo(true);
+              if (candidate.is_source) groupsModal.current.present();
             }}
           >
             {candidate.is_source ? "Previously Saved" : "Not saved"}
@@ -67,7 +63,7 @@ const ScanningCardBase = ({
         </div>
         <PinnedAnnotations
           source={candidate}
-          onButtonClick={() => modal.current?.present()}
+          onButtonClick={() => annotationsModal.current?.present()}
           pinnedAnnotationIds={pinnedAnnotations}
         />
         <div className="plot-container">
@@ -83,28 +79,9 @@ const ScanningCardBase = ({
         <FollowupRequests source={candidate} requestType={"forced_photometry"}/>
       </div>
       <ScanningCardSkeleton visible={!isInView} />
-      {/* Saved groups modal */}
-      <IonModal isOpen={showGroupsSaveTo} onDidDismiss={() => setShowGroupsSaveTo(false)}>
-        <IonHeader>
-          <IonToolbar>
-            <IonTitle>Saved Groups</IonTitle>
-            <IonButtons slot="end">
-              <IonButton onClick={() => setShowGroupsSaveTo(false)}>Close</IonButton>
-            </IonButtons>
-          </IonToolbar>
-        </IonHeader>
-        <IonContent className="ion-padding">
-          {candidate.saved_groups?.length ? (
-            candidate.saved_groups.map((group) => (
-              <IonChip key={group.name} color="secondary">
-                {group.name}
-              </IonChip>
-            ))
-          ) : (
-            <p>No saved groups</p>
-          )}
-        </IonContent>
-      </IonModal>
+      {/* Modals */}
+      <GroupsModal title="Saved Groups" groups={candidate.saved_groups} modal={groupsModal} />
+      <AnnotationsViewerModal source={candidate} modal={annotationsModal}/>
     </div>
   );
 };
